feat: implement modules part of task
