+++ src/modules/esm.mjs
@@ -1,0 +1,35 @@
+import * as path from "node:path";
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'http';
+import { readFile } from "node:fs/promises";
+import './files/c.cjs';
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = await readFile(path.join(import.meta.dirname, './files/a.json'));
+} else {
+    unknownObject = await readFile(path.join(import.meta.dirname,'./files/b.json'));
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${import.meta.filename}`);
+console.log(`Path to current directory is ${import.meta.dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(JSON.parse(unknownObject));
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
